Add search functionality
Pressing `/` switches into filtering mode and lets the user type a
search string. The filtering is done by looking for the characters
of the search string in the node name, maintaining the order, but
not requiring the characters to be consequtive.

For example, `ob` will match `foo-bar`.

The filtered list also includes the currently active exit node, and the
node that has a cursor on it, regardless of the name.